by D 数据修改json

+++ src/com/casic/datadriver/model/data/PrivateData.java
@@ -29,9 +29,9 @@ public class PrivateData extends BaseModel{
 
     private Long ddDataCreatorId;
 
-    private Long ddDataSenMin;
+    private Double ddDataSenMin;
 
-    private Long ddDataSenMax;
+    private Double ddDataSenMax;
 
     private Long ddDataParentId;
 
@@ -157,19 +157,19 @@ public class PrivateData extends BaseModel{
         this.ddDataCreatorId = ddDataCreatorId;
     }
 
-    public Long getDdDataSenMin() {
+    public Double getDdDataSenMin() {
         return ddDataSenMin;
     }
 
-    public void setDdDataSenMin(Long ddDataSenMin) {
+    public void setDdDataSenMin(Double ddDataSenMin) {
         this.ddDataSenMin = ddDataSenMin;
     }
 
-    public Long getDdDataSenMax() {
+    public Double getDdDataSenMax() {
         return ddDataSenMax;
     }
 
-    public void setDdDataSenMax(Long ddDataSenMax) {
+    public void setDdDataSenMax(Double ddDataSenMax) {
         this.ddDataSenMax = ddDataSenMax;
     }
